References

EvolutionaryModelingTools.Reaction — Type

Reaction{C,U}

Contain a "calculate" function and an "update" function. The "calculate" function calculates the a "rate" of reaction with the system state, which determines the probability weight of the reaction be selected, and the "update" function updates the system state, when this reaction were selected randomly.

EvolutionaryModelingTools.@cfunc — Macro

@cfunc ex

Define a "calculation" function with an "adapter" methods used to parse args from model. "calculate" functions take arguments from system state and calculate "rate"s determining the probability weight of the reaction be selected.

Example

For function definition:

@cfunc @inline growth_c(r, x::Vector) = r * x # function to calculate "growth rate"

This macro creates two methods, an "adapter" method

growth_c(args::NamedTuple) = growth_c(args.r, args.x)

and the origin method

@inline growth_c(r, x::Vector) = r * x

Info

For function definition with other macros, put those macros after this macro.

Warning

The argument name t is reserved for time. Don't use those variable names for other usage

EvolutionaryModelingTools.@ufunc — Macro

@ufunc ex

Define a "update" function with an "adapter" methods used to parse args from model. "update" functions take arguments from system state and update system state.

Example

For function definition:

@ufunc Base.@propagate_inbounds growth_u!(ind::CartesianIndex{1}, x) = x[ind] += 1

This macro creates two methods, an "adapter" method

growth_u!(ind, args::NamedTuple) = growth_u(ind, args.x)

and the origin method

Base.@propagate_inbounds growth_u!(ind::CartesianIndex{1}, x) = x[ind] += 1

Info

For function definition with other macros, put those macros after this macro.

Warning

The argument name t is reserved for time, arguments name rng is reserved for random number generator, and the argument name ind is preserved for index of "reaction". Don't use those variable names for other usage

EvolutionaryModelingTools.@reaction — Macro

@reaction name ex

Define a Reaction with the given name and ex.

@reaction growth begin
    r * x # "calculate" expression
    begin
        i = ind[1]
        x[i] += 1 # "update" expression
    end
end

will create a Reaction with a "calculation" function:

@cfunc Base.@propagate_inbounds growth_c(r, x) = r * x

and an "update" function:

@ufunc Base.@propagate_inbounds growth_u!(ind, x) = begin
    i = ind[1]
    x[i] += 1
end

where arguments of functions were collected from given expression automatically.

Note

If there are global variables, this macro may can not collect arguments correctly Especially, for functions which accept function and types as arguments, those functions and types may also be collect as an arguments. Thus, these variables must be marked as global variables by global before use them, even functions and types. Besides, type annotation expressions like x::Vector{Int}, types Vector and Int will not be collected. Avoid to defined your reaction with a type arguments for type annotation.

Warning

The expression follow the same name preserve rule as @cfunc and @ufunc, don't use those variable names for other usage.

This macro cannot parse macro expression. Thus in some cases, arguments of expression with macro may not be collected correctly. To avoid this, define reaction with anonymous functions may helpfully:

@reaction reaction begin
    (A, B) -> @einsum C[i, k] := A[i, j] * B[j, k] # where C, i, j, k should not be collected
    begin
        # do something
    end
end

in this case, only arguments of anonymous function (c, A and B) will be collected.

EvolutionaryModelingTools.gillespie — Function

gillespie([hook!, rng::AbstractRNG,] c, ps::NamedTuple, rs::Tuple)

Simulate the system using the Gillespie algorithm with the given parameters, and return a tuple of updated ps, end time and terminate state. More about terminate state, see gillespie!.

Arguments

hook!: a function with similar arguments to "update" functions and recommended to created with @ufunc macro; unlike "update" functions, hook will be called after each reaction and should return a terminate state used to terminate the simulation if it is not :finnish.

rng: a random number generator for generate random numbers;

c: a ContinuousClock, a end time or a tuple of a begin and a end time;

ps: a NamedTuple contains state, parameters even args used by hook! of the system;

rs: a tuple contains reactions, all parameters required by reactions must be in ps with same name.

EvolutionaryModelingTools.gillespie! — Function

gillespie!(hook!, rng::AbstractRNG, c::ContinuousClock, ps::NamedTuple, rs::Tuple)

Simulate the system using the Gillespie algorithm with the given parameters, and return a tuple of end time and the terminate state :finnish, :zero or any other state returns by the hook!. The terminate state :finnish means that simulation reach to the end time, and :zero means the simulation break because the total "reaction rate" is zero, besides, hook! should return a symbol terminate state like :break. if the return value of hook! is not :finnish, the simulation will be terminated. The clock c and parameters ps will be updated during the simulation.

Arguments

hook!: a function with similar arguments to "update" functions, and it's recommended to create hook! with @ufunc macro; unlike "update" functions, hook will be called after each reaction and should return a terminate state used to terminate the simulation if it is not :finnish.

rng: a random number generator for generate random numbers;

c: a clock recording time, which must be the reference clock of recorded variables in ps;

ps: a NamedTuple contains state, parameters even args used by hook! of the system;

rs: a tuple contains reactions, all parameters required by reactions must be in ps with same name.